+++ ui/src/core/misc/utils.cpp
@@ -2,6 +2,7 @@
 #include "core/misc/utils.hpp"
 
 #include <Qt>
+#include <QThread>
 
 std::optional<uint8_t> keyToHex(int const key) {
     if (key >= Qt::Key_0 && key <= Qt::Key_9) {
@@ -19,4 +20,8 @@ uint8_t replaceNibble(uint8_t value, uint8_t nibble, bool highNibble) {
     } else {
         return (value & 0xF0) | (nibble);
     }
+}
+
+bool objectInCurrentThread(QObject const& obj) {
+    return obj.thread() == QThread::currentThread();
 }